flask/api/requirements.txt
amqp==2.3.2
aniso8601==3.0.2
appnope==0.1.0
backcall==0.1.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.4.16
click==6.7
decorator==4.3.0
Flask==1.0.2
Flask-RESTful==0.3.6
ipdb==0.9.4
ipython==6.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
pytz==2018.5
redis==2.10.6
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
vine==1.1.4
wcwidth==0.1.7
Werkzeug==0.14.1
